co-locate everything
