+++ js/src/instrumentation/plugins/openrouter-plugin.ts
@@ -10,24 +10,17 @@ import { SpanTypeAttribute, isObject } from "../../../util/index";
 import { withCurrent } from "../../logger";
 import type { Span } from "../../logger";
 import { getCurrentUnixTimestamp } from "../../util";
-import { patchOpenRouterCallModelRequestTools } from "../../openrouter-tool-wrapping";
-import {
-  buildOpenRouterEmbeddingMetadata,
-  buildOpenRouterMetadata,
-  extractOpenRouterCallModelInput,
-  extractOpenRouterCallModelMetadata,
-  extractOpenRouterResponseMetadata,
-  extractOpenRouterResponseOutput,
-  sanitizeOpenRouterLoggedValue,
-} from "../../openrouter-logging";
-import { parseOpenRouterMetricsFromUsage } from "../../openrouter-utils";
+import { zodToJsonSchema } from "../../zod/utils";
 import { openRouterChannels } from "./openrouter-channels";
 import type {
   OpenRouterChatChoice,
   OpenRouterChatCompletionChunk,
   OpenRouterCallModelRequest,
+  OpenRouterEmbeddingResponse,
   OpenRouterResponse,
   OpenRouterResponseStreamEvent,
+  OpenRouterTool,
+  OpenRouterToolTurnContext,
 } from "../../vendor-sdk-types/openrouter";
 
 export class OpenRouterPlugin extends BasePlugin {
@@ -308,6 +301,450 @@ function getOpenRouterCallModelRequestArg(
   return isObject(firstObjectArg)
     ? (firstObjectArg as OpenRouterCallModelRequest)
     : undefined;
+}
+
+const TOKEN_NAME_MAP: Record<string, string> = {
+  promptTokens: "prompt_tokens",
+  inputTokens: "prompt_tokens",
+  completionTokens: "completion_tokens",
+  outputTokens: "completion_tokens",
+  totalTokens: "tokens",
+  prompt_tokens: "prompt_tokens",
+  input_tokens: "prompt_tokens",
+  completion_tokens: "completion_tokens",
+  output_tokens: "completion_tokens",
+  total_tokens: "tokens",
+};
+
+const TOKEN_DETAIL_PREFIX_MAP: Record<string, string> = {
+  promptTokensDetails: "prompt",
+  inputTokensDetails: "prompt",
+  completionTokensDetails: "completion",
+  outputTokensDetails: "completion",
+  costDetails: "cost",
+  prompt_tokens_details: "prompt",
+  input_tokens_details: "prompt",
+  completion_tokens_details: "completion",
+  output_tokens_details: "completion",
+  cost_details: "cost",
+};
+
+function camelToSnake(value: string): string {
+  return value.replace(/[A-Z]/g, (match) => `_${match.toLowerCase()}`);
+}
+
+function parseOpenRouterMetricsFromUsage(
+  usage: unknown,
+): Record<string, number> {
+  if (!isObject(usage)) {
+    return {};
+  }
+
+  const metrics: Record<string, number> = {};
+
+  for (const [name, value] of Object.entries(usage)) {
+    if (typeof value === "number") {
+      metrics[TOKEN_NAME_MAP[name] || camelToSnake(name)] = value;
+      continue;
+    }
+
+    if (!isObject(value)) {
+      continue;
+    }
+
+    const prefix = TOKEN_DETAIL_PREFIX_MAP[name];
+    if (!prefix) {
+      continue;
+    }
+
+    for (const [nestedName, nestedValue] of Object.entries(value)) {
+      if (typeof nestedValue !== "number") {
+        continue;
+      }
+
+      metrics[`${prefix}_${camelToSnake(nestedName)}`] = nestedValue;
+    }
+  }
+
+  return metrics;
+}
+
+function extractOpenRouterUsageMetadata(
+  usage: unknown,
+): Record<string, unknown> | undefined {
+  if (!isObject(usage)) {
+    return undefined;
+  }
+
+  const metadata: Record<string, unknown> = {};
+
+  if (typeof usage.isByok === "boolean") {
+    metadata.is_byok = usage.isByok;
+  } else if (typeof usage.is_byok === "boolean") {
+    metadata.is_byok = usage.is_byok;
+  }
+
+  return Object.keys(metadata).length > 0 ? metadata : undefined;
+}
+
+const OMITTED_OPENROUTER_KEYS = new Set([
+  "execute",
+  "render",
+  "nextTurnParams",
+  "requireApproval",
+]);
+
+function parseOpenRouterModelString(model: unknown): {
+  model: unknown;
+  provider?: string;
+} {
+  if (typeof model !== "string") {
+    return { model };
+  }
+
+  const slashIndex = model.indexOf("/");
+  if (slashIndex > 0 && slashIndex < model.length - 1) {
+    return {
+      provider: model.substring(0, slashIndex),
+      model: model.substring(slashIndex + 1),
+    };
+  }
+
+  return { model };
+}
+
+function isZodSchema(value: unknown): boolean {
+  return (
+    value != null &&
+    typeof value === "object" &&
+    "_def" in value &&
+    typeof (value as { _def?: unknown })._def === "object"
+  );
+}
+
+function serializeZodSchema(schema: unknown): Record<string, unknown> {
+  try {
+    return zodToJsonSchema(schema as any) as Record<string, unknown>;
+  } catch {
+    return {
+      type: "object",
+      description: "Zod schema (conversion failed)",
+    };
+  }
+}
+
+function serializeOpenRouterTool(tool: OpenRouterTool): OpenRouterTool {
+  if (!isObject(tool)) {
+    return tool;
+  }
+
+  const serialized: Record<string, unknown> = {};
+  for (const [key, value] of Object.entries(tool)) {
+    if (OMITTED_OPENROUTER_KEYS.has(key)) {
+      continue;
+    }
+
+    if (key === "function" && isObject(value)) {
+      serialized.function = sanitizeOpenRouterLoggedValue(value);
+      continue;
+    }
+
+    serialized[key] = sanitizeOpenRouterLoggedValue(value);
+  }
+
+  return serialized as OpenRouterTool;
+}
+
+function serializeOpenRouterToolsForLogging(
+  tools: readonly OpenRouterTool[] | undefined,
+): OpenRouterTool[] | undefined {
+  if (!Array.isArray(tools)) {
+    return undefined;
+  }
+
+  return tools.map((tool) => serializeOpenRouterTool(tool));
+}
+
+function sanitizeOpenRouterLoggedValue(value: unknown): unknown {
+  if (isZodSchema(value)) {
+    return serializeZodSchema(value);
+  }
+
+  if (typeof value === "function") {
+    return "[Function]";
+  }
+
+  if (Array.isArray(value)) {
+    return value.map((entry) => sanitizeOpenRouterLoggedValue(entry));
+  }
+
+  if (!isObject(value)) {
+    return value;
+  }
+
+  const sanitized: Record<string, unknown> = {};
+  for (const [key, entry] of Object.entries(value)) {
+    if (OMITTED_OPENROUTER_KEYS.has(key)) {
+      continue;
+    }
+
+    if (key === "tools" && Array.isArray(entry)) {
+      sanitized.tools = serializeOpenRouterToolsForLogging(entry);
+      continue;
+    }
+
+    sanitized[key] = sanitizeOpenRouterLoggedValue(entry);
+  }
+
+  return sanitized;
+}
+
+function buildOpenRouterMetadata(
+  metadata: Record<string, unknown>,
+  httpReferer: unknown,
+  xTitle: unknown,
+): Record<string, unknown> {
+  const sanitized = sanitizeOpenRouterLoggedValue(metadata);
+  const metadataRecord = isObject(sanitized) ? sanitized : {};
+  const { model, provider: providerRouting, ...rest } = metadataRecord;
+  const normalizedModel = parseOpenRouterModelString(model);
+
+  return {
+    ...rest,
+    ...(normalizedModel.model !== undefined
+      ? { model: normalizedModel.model }
+      : {}),
+    ...(providerRouting !== undefined ? { providerRouting } : {}),
+    ...(httpReferer !== undefined ? { httpReferer } : {}),
+    ...(xTitle !== undefined ? { xTitle } : {}),
+    provider: normalizedModel.provider || "openrouter",
+  };
+}
+
+function buildOpenRouterEmbeddingMetadata(
+  metadata: Record<string, unknown>,
+  httpReferer: unknown,
+  xTitle: unknown,
+): Record<string, unknown> {
+  const normalized = buildOpenRouterMetadata(metadata, httpReferer, xTitle);
+  return typeof normalized.model === "string"
+    ? {
+        ...normalized,
+        embedding_model: normalized.model,
+      }
+    : normalized;
+}
+
+function extractOpenRouterCallModelInput(
+  request: OpenRouterCallModelRequest,
+): unknown {
+  return isObject(request) && "input" in request
+    ? sanitizeOpenRouterLoggedValue(request.input)
+    : undefined;
+}
+
+function extractOpenRouterCallModelMetadata(
+  request: OpenRouterCallModelRequest,
+): Record<string, unknown> {
+  if (!isObject(request)) {
+    return { provider: "openrouter" };
+  }
+
+  const { input: _input, ...metadata } = request;
+  return buildOpenRouterMetadata(metadata, undefined, undefined);
+}
+
+function extractOpenRouterResponseMetadata(
+  result: OpenRouterResponse | OpenRouterEmbeddingResponse | undefined,
+): Record<string, unknown> | undefined {
+  if (!isObject(result)) {
+    return undefined;
+  }
+
+  const { output: _output, data: _data, usage, ...metadata } = result;
+  const sanitized = sanitizeOpenRouterLoggedValue(metadata);
+  const metadataRecord = isObject(sanitized) ? sanitized : {};
+  const { model, provider, ...rest } = metadataRecord;
+  const normalizedModel = parseOpenRouterModelString(model);
+  const normalizedProvider =
+    (typeof provider === "string" ? provider : undefined) ||
+    normalizedModel.provider;
+  const usageMetadata = extractOpenRouterUsageMetadata(usage);
+  const combined = {
+    ...rest,
+    ...(normalizedModel.model !== undefined
+      ? { model: normalizedModel.model }
+      : {}),
+    ...(usageMetadata || {}),
+    ...(normalizedProvider !== undefined
+      ? { provider: normalizedProvider }
+      : {}),
+  };
+
+  return Object.keys(combined).length > 0 ? combined : undefined;
+}
+
+function extractOpenRouterResponseOutput(
+  response: Record<string, unknown> | undefined,
+  fallbackOutput?: unknown,
+): unknown {
+  if (
+    isObject(response) &&
+    "output" in response &&
+    response.output !== undefined
+  ) {
+    return sanitizeOpenRouterLoggedValue(response.output);
+  }
+
+  if (fallbackOutput !== undefined) {
+    return sanitizeOpenRouterLoggedValue(fallbackOutput);
+  }
+
+  return undefined;
+}
+
+const OPENROUTER_WRAPPED_TOOL = Symbol("braintrust.openrouter.wrappedTool");
+
+type OpenRouterToolTraceContext = ChannelMessage<
+  typeof openRouterChannels.toolExecute
+>;
+
+type WrappedOpenRouterTool = OpenRouterTool & {
+  [OPENROUTER_WRAPPED_TOOL]?: true;
+};
+
+function patchOpenRouterCallModelRequestTools(
+  request: OpenRouterCallModelRequest,
+): (() => void) | undefined {
+  if (!Array.isArray(request.tools) || request.tools.length === 0) {
+    return undefined;
+  }
+
+  const originalTools = request.tools;
+  const wrappedTools = originalTools.map((tool) => wrapOpenRouterTool(tool));
+  const didPatch = wrappedTools.some(
+    (tool, index) => tool !== originalTools[index],
+  );
+  if (!didPatch) {
+    return undefined;
+  }
+
+  (request as { tools?: readonly OpenRouterTool[] }).tools = wrappedTools;
+  return () => {
+    (request as { tools?: readonly OpenRouterTool[] }).tools = originalTools;
+  };
+}
+
+function wrapOpenRouterTool(tool: OpenRouterTool): OpenRouterTool {
+  if (
+    isWrappedTool(tool) ||
+    !tool.function ||
+    typeof tool.function !== "object" ||
+    typeof tool.function.execute !== "function"
+  ) {
+    return tool;
+  }
+
+  const toolName = tool.function.name || "tool";
+  const originalExecute = tool.function.execute;
+  const wrappedTool: WrappedOpenRouterTool = {
+    ...tool,
+    function: {
+      ...tool.function,
+      execute(this: unknown, ...args: unknown[]) {
+        return traceToolExecution({
+          args,
+          execute: () => Reflect.apply(originalExecute, this, args),
+          toolCallId: getToolCallId(args[1]),
+          toolName,
+        });
+      },
+    },
+  };
+
+  Object.defineProperty(wrappedTool, OPENROUTER_WRAPPED_TOOL, {
+    value: true,
+    enumerable: false,
+    configurable: false,
+  });
+
+  return wrappedTool;
+}
+
+function isWrappedTool(tool: OpenRouterTool): boolean {
+  return Boolean((tool as WrappedOpenRouterTool)[OPENROUTER_WRAPPED_TOOL]);
+}
+
+function traceToolExecution(args: {
+  args: unknown[];
+  execute: () => unknown;
+  toolCallId?: string;
+  toolName: string;
+}): unknown {
+  const tracingChannel = openRouterChannels.toolExecute.tracingChannel();
+  const input = args.args.length > 0 ? args.args[0] : undefined;
+  const event: OpenRouterToolTraceContext = {
+    arguments: [input],
+    span_info: {
+      name: args.toolName,
+    },
+    toolCallId: args.toolCallId,
+    toolName: args.toolName,
+  };
+
+  tracingChannel.start!.publish(event);
+
+  try {
+    const result = args.execute();
+    return publishToolResult(tracingChannel, event, result);
+  } catch (error) {
+    event.error = normalizeError(error);
+    tracingChannel.error!.publish(event);
+    throw error;
+  }
+}
+
+function publishToolResult(
+  tracingChannel: ReturnType<
+    typeof openRouterChannels.toolExecute.tracingChannel
+  >,
+  event: OpenRouterToolTraceContext,
+  result: unknown,
+): unknown {
+  if (isPromiseLike(result)) {
+    return result.then(
+      (resolved) => {
+        event.result = resolved;
+        tracingChannel.asyncEnd!.publish(event);
+        return resolved;
+      },
+      (error) => {
+        event.error = normalizeError(error);
+        tracingChannel.error!.publish(event);
+        throw error;
+      },
+    );
+  }
+
+  event.result = result;
+  tracingChannel.asyncEnd!.publish(event);
+  return result;
+}
+
+function getToolCallId(context: unknown): string | undefined {
+  const toolContext = context as OpenRouterToolTurnContext | undefined;
+  return typeof toolContext?.toolCall?.id === "string"
+    ? toolContext.toolCall.id
+    : undefined;
+}
+
+function isPromiseLike(value: unknown): value is PromiseLike<unknown> {
+  return (
+    !!value &&
+    (typeof value === "object" || typeof value === "function") &&
+    "then" in value &&
+    typeof value.then === "function"
+  );
 }
 
 export function aggregateOpenRouterChatChunks(
